[backend] NLQ LLM implem
