+++ opencti-platform/opencti-graphql/src/database/ai-llm.ts
@@ -1,6 +1,8 @@
 import { Mistral } from '@mistralai/mistralai';
 import OpenAI from 'openai';
 import type { ChatCompletionStreamRequest } from '@mistralai/mistralai/models/components';
+import { ChatOpenAI } from '@langchain/openai';
+import type { ChatPromptValueInterface } from '@langchain/core/prompt_values';
 import conf, { BUS_TOPICS, logApp } from '../config/conf';
 import { isEmptyField } from './utils';
 import { notify } from './redis';
@@ -17,6 +19,7 @@ const AI_MODEL = conf.get('ai:model');
 const AI_MAX_TOKENS = conf.get('ai:max_tokens');
 
 let client: Mistral | OpenAI | null = null;
+let llm = null;
 if (AI_ENABLED && AI_TOKEN) {
   switch (AI_TYPE) {
     case 'mistralai':
@@ -29,6 +32,14 @@ if (AI_ENABLED && AI_TOKEN) {
           group: (label) => logApp.info(`[AI] group ${label} start.`),
           groupEnd: () => logApp.info('[AI] group end.'),
         } */
+      });
+      llm = new ChatOpenAI({
+        modelName: 'mistral',
+        apiKey: AI_TOKEN,
+        temperature: 0,
+        configuration: {
+          baseURL: 'https://ai.filigran.io/v1',
+        },
       });
       break;
     case 'openai':
@@ -129,4 +140,11 @@ export const queryAi = async (busId: string | null, developerMessage: string | n
     default:
       throw UnsupportedError('Not supported AI type', { type: AI_TYPE });
   }
+};
+
+export const queryNLQAi = async (promptValue: ChatPromptValueInterface) => {
+  if (!llm) {
+    throw UnsupportedError('Incorrect AI configuration', { enabled: AI_ENABLED, type: AI_TYPE, endpoint: AI_ENDPOINT, model: AI_MODEL });
+  }
+  return llm.invoke(promptValue);
 };
