--- opencti-platform/opencti-graphql/src/modules/ai/ai-utils.ts
@@ -1,451 +1,0 @@
-export const jsonFiltersSchema = {
-  response_format: {
-    type: 'json_schema',
-    json_schema: {
-      schema: {
-        $defs: {
-          FilterItem: {
-            description: 'Represents a filter item that handles (both relationship) and entity filtering scenarios.\n\nExample:\n{\n  "key": "entity_type",\n  "operator": "eq",\n  "values": [\n      "Threat-Actor-Group",\n      "Threat-Actor-Individual"\n      OR\n      { "key": "relationship_type", "values": ["targets"] }\n  ],\n  "mode": "or"\n}',
-            properties: {
-              key: {
-                anyOf: [
-                  {
-                    $ref: '#/$defs/KeyTypeEnum'
-                  },
-                  {
-                    type: 'null'
-                  }
-                ],
-                description: "The main filter key (e.g. 'entity_type', 'regardingOf', 'creator_id', etc.)."
-              },
-              operator: {
-                default: 'eq',
-                description: "The logic operator used for the filter (e.g., 'eq', 'lte', 'gte', ...).",
-                enum: [
-                  'eq',
-                  'not_eq',
-                  'gt',
-                  'gte',
-                  'lt',
-                  'lte',
-                  'nil',
-                  'not_nil',
-                  'starts_with',
-                  'not_starts_with',
-                  'ends_with',
-                  'not_ends_with',
-                  'contains',
-                  'not_contains',
-                  'search'
-                ],
-                title: 'OperatorEnum',
-                type: 'string'
-              },
-              values: {
-                description: 'A list of values. Each value can be a simple string or a nested FilterValueContainer.\nExamples:\n  - ["Intrusion-Set", "Threat-Actor-Group"]\n  - [{"key": "id", "values": ["some-uuid"]}]',
-                items: {
-                  anyOf: [
-                    {
-                      type: 'string'
-                    },
-                    {
-                      $ref: '#/$defs/FilterValueContainer'
-                    }
-                  ]
-                },
-                title: 'Values',
-                type: 'array'
-              },
-              mode: {
-                default: 'or',
-                description: "Filter combination mode for this item\u2019s values (often 'or').",
-                title: 'Mode',
-                type: 'string'
-              }
-            },
-            title: 'FilterItem',
-            type: 'object',
-            additionalProperties: false,
-            required: [
-              'key',
-              'operator',
-              'values',
-              'mode'
-            ]
-          },
-          FilterValueContainer: {
-            description: 'A container for holding a filter key and its associated values.\n\nAttributes:\n    key (str): The identifier for the filter.\n    values (list[str]): A list of values corresponding to the filter key.',
-            properties: {
-              key: {
-                anyOf: [
-                  {
-                    enum: [
-                      'relationship_type',
-                      'id'
-                    ],
-                    type: 'string'
-                  },
-                  {
-                    type: 'null'
-                  }
-                ],
-                description: "Sub-key used in a nested filter condition (e.g. 'id', 'relationship_type').",
-                title: 'Key'
-              },
-              values: {
-                description: 'A list of string values for the given sub-key.',
-                items: {
-                  type: 'string'
-                },
-                title: 'Values',
-                type: 'array'
-              }
-            },
-            title: 'FilterValueContainer',
-            type: 'object',
-            additionalProperties: false,
-            required: [
-              'key',
-              'values'
-            ]
-          },
-          KeyTypeEnum: {
-            description: "Possible values for the 'key' field when building OpenCTI filters.",
-            enum: [
-              'entity_type',
-              'created_at',
-              'updated_at',
-              'creator_id',
-              'opinions_metrics.mean',
-              'opinions_metrics.max',
-              'opinions_metrics.min',
-              'opinions_metrics.total',
-              'createdBy',
-              'objectMarking',
-              'objectLabel',
-              'externalReferences',
-              'regardingOf',
-              'computed_reliability',
-              'objects',
-              'workflow_id',
-              'created',
-              'confidence',
-              'name',
-              'description',
-              'x_mitre_platforms',
-              'x_mitre_permissions_required',
-              'x_mitre_detection',
-              'killChainPhases',
-              'alias',
-              'first_seen',
-              'last_seen',
-              'objective',
-              'objectOrganization',
-              'attribute_abstract',
-              'content',
-              'note_types',
-              'likelihood',
-              'first_observed',
-              'last_observed',
-              'number_observed',
-              'explanation',
-              'opinion',
-              'report_types',
-              'published',
-              'objectAssignee',
-              'objectParticipant',
-              'x_mitre_id',
-              'x_opencti_threat_hunting',
-              'x_opencti_log_sources',
-              'contact_information',
-              'infrastructure_types',
-              'goals',
-              'resource_level',
-              'primary_motivation',
-              'secondary_motivations',
-              'postal_code',
-              'street_address',
-              'malware_types',
-              'is_family',
-              'architecture_execution_envs',
-              'implementation_languages',
-              'capabilities',
-              'samples',
-              'operatingSystems',
-              'threat_actor_types',
-              'roles',
-              'sophistication',
-              'tool_types',
-              'tool_version',
-              'x_opencti_cvss_base_score',
-              'x_opencti_cvss_base_severity',
-              'x_opencti_cvss_attack_vector',
-              'x_opencti_cvss_integrity_impact',
-              'x_opencti_cvss_availability_impact',
-              'x_opencti_cvss_confidentiality_impact',
-              'x_opencti_cisa_kev',
-              'x_opencti_epss_score',
-              'x_opencti_epss_percentile',
-              'incident_type',
-              'severity',
-              'source',
-              'channel_types',
-              'event_types',
-              'start_time',
-              'stop_time',
-              'context',
-              'narrative_types',
-              'dataSource',
-              'collection_layers',
-              'due_date',
-              'priority',
-              'response_types',
-              'information_types',
-              'takedown_types',
-              'rating',
-              'product',
-              'version',
-              'configuration_version',
-              'modules',
-              'analysis_engine_version',
-              'analysis_definition_version',
-              'submitted',
-              'analysis_started',
-              'analysis_ended',
-              'result',
-              'hostVm',
-              'operatingSystem',
-              'installedSoftware',
-              'analysisSco',
-              'analysisSample',
-              'revoked',
-              'personal_motivations',
-              'date_of_birth',
-              'gender',
-              'job_title',
-              'marital_status',
-              'eye_color',
-              'hair_color',
-              'bornIn',
-              'ethnicity',
-              'pattern_type',
-              'pattern',
-              'indicator_types',
-              'valid_from',
-              'valid_until',
-              'x_opencti_score',
-              'x_opencti_detection',
-              'x_opencti_main_observable_type',
-              'x_opencti_organization_type',
-              'x_opencti_description',
-              'number',
-              'rir',
-              'path',
-              'path_enc',
-              'ctime',
-              'mtime',
-              'atime',
-              'containsObservable',
-              'value',
-              'resolvesTo',
-              'display_name',
-              'belongsTo',
-              'is_multipart',
-              'attribute_date',
-              'content_type',
-              'message_id',
-              'subject',
-              'received_lines',
-              'body',
-              'emailFrom',
-              'emailSender',
-              'emailTo',
-              'emailCc',
-              'emailBcc',
-              'bodyMultipart',
-              'rawEmail',
-              'content_disposition',
-              'bodyRaw',
-              'hashes.MD5',
-              'hashes.SHA-1',
-              'hashes.SHA-256',
-              'hashes.SHA-512',
-              'hashes.SSDEEP',
-              'mime_type',
-              'payload_bin',
-              'url',
-              'encryption_algorithm',
-              'decryption_key',
-              'x_opencti_additional_names',
-              'extensions',
-              'size',
-              'name_enc',
-              'magic_number_hex',
-              'parentDirectory',
-              'obsContent',
-              'is_self_signed',
-              'serial_number',
-              'signature_algorithm',
-              'issuer',
-              'validity_not_before',
-              'validity_not_after',
-              'subject_public_key_algorithm',
-              'subject_public_key_modulus',
-              'subject_public_key_exponent',
-              'basic_constraints',
-              'name_constraints',
-              'policy_constraints',
-              'key_usage',
-              'extended_key_usage',
-              'subject_key_identifier',
-              'authority_key_identifier',
-              'subject_alternative_name',
-              'issuer_alternative_name',
-              'subject_directory_attributes',
-              'crl_distribution_points',
-              'inhibit_any_policy',
-              'private_key_usage_period_not_before',
-              'private_key_usage_period_not_after',
-              'certificate_policies',
-              'policy_mappings',
-              'start',
-              'end',
-              'is_active',
-              'src_port',
-              'dst_port',
-              'protocols',
-              'src_byte_count',
-              'dst_byte_count',
-              'src_packets',
-              'dst_packets',
-              'networkSrc',
-              'networkDst',
-              'srcPayload',
-              'dstPayload',
-              'networkEncapsulates',
-              'encapsulatedBy',
-              'is_hidden',
-              'pid',
-              'created_time',
-              'cwd',
-              'command_line',
-              'environment_variables',
-              'aslr_enabled',
-              'dep_enabled',
-              'owner_sid',
-              'window_title',
-              'integrity_level',
-              'service_name',
-              'descriptions',
-              'group_name',
-              'start_type',
-              'service_type',
-              'service_status',
-              'openedConnections',
-              'creatorUser',
-              'processImage',
-              'processParent',
-              'processChild',
-              'serviceDlls',
-              'cpe',
-              'swid',
-              'languages',
-              'vendor',
-              'user_id',
-              'credential',
-              'account_login',
-              'account_type',
-              'is_service_account',
-              'is_privileged',
-              'can_escalate_privs',
-              'is_disabled',
-              'account_created',
-              'account_expires',
-              'credential_last_changed',
-              'account_first_login',
-              'account_last_login',
-              'attribute_key',
-              'modified_time',
-              'number_of_subkeys',
-              'winRegValues',
-              'data',
-              'data_type',
-              'iban',
-              'bic',
-              'account_number',
-              'card_number',
-              'expiration_date',
-              'cvv',
-              'holder_name',
-              'title',
-              'media_category',
-              'publication_date',
-              'persona_name',
-              'persona_type',
-              'relationship_type',
-              'id'
-            ],
-            title: 'KeyTypeEnum',
-            type: 'string'
-          },
-          OperatorEnum: {
-            enum: [
-              'eq',
-              'not_eq',
-              'gt',
-              'gte',
-              'lt',
-              'lte',
-              'nil',
-              'not_nil',
-              'starts_with',
-              'not_starts_with',
-              'ends_with',
-              'not_ends_with',
-              'contains',
-              'not_contains',
-              'search'
-            ],
-            title: 'OperatorEnum',
-            type: 'string'
-          }
-        },
-        description: 'Represents the output structure for OpenCTI filters.\n\nExample:\n{\n  "mode": "and",\n  "filters": [ { ... }, { ... } ],\n  "filterGroups": []\n}',
-        properties: {
-          mode: {
-            default: 'and',
-            description: "Logical operator combining all filter items at top level ('and' or 'or').",
-            title: 'Mode',
-            type: 'string'
-          },
-          filters: {
-            description: "The list of FilterItems (conditions). Each FilterItem can have its own 'key', 'operator', 'values', etc.",
-            items: {
-              $ref: '#/$defs/FilterItem'
-            },
-            title: 'Filters',
-            type: 'array'
-          },
-          filterGroups: {
-            description: 'Allows for nested or grouped filters if needed. Often empty in the examples.',
-            items: {},
-            title: 'Filtergroups',
-            type: 'array'
-          }
-        },
-        title: 'OpenCTIFiltersOutput',
-        type: 'object',
-        additionalProperties: false,
-        required: [
-          'mode',
-          'filters',
-          'filterGroups'
-        ]
-      },
-      name: 'OpenCTIFiltersOutput',
-      strict: true
-    }
-  },
-  stream: false,
-  temperature: 0.0
-};
